consolidate Course API calls to one util file
