--- client/src/utils/deleteCourse/index.js
@@ -1,25 +1,0 @@
-import { apiUrl, tokenKey } from "../config"
-
-export const deleteCourse = (courseName) => {
-    const url = `${apiUrl}/tutor/courses`
-    const options = {
-        method: 'DELETE',
-        headers: {
-            'Content-Type': 'application/json',
-            'authorization': `Bearer: ${localStorage.getItem(tokenKey)}`
-        },
-        body: JSON.stringify({ courseName: courseName }),
-    }
-
-    return new Promise(async (resolve, reject) => {
-        try {
-            const data = await fetch(url, options).then(res => res.json())
-            if (!data) return reject(null)
-            resolve(data)
-
-        } catch (error) {
-            console.error(error)
-            reject(null)
-        }
-    })
-}
