+++ client/src/utils/index.js
@@ -2,8 +2,7 @@ import { signup } from "./signup";
 import { loginWithToken } from "./loginWithToken";
 import { loginWithPassword } from "./loginWithPassword";
 import { logoutTutor } from "./logout";
-import { addCourse } from "./addCourse";
-import { deleteCourse } from "./deleteCourse";
+import { addCourse, deleteCourse } from "./Course";
 
 export {
     signup,
